CMake: EXPORT targets/configs (Finally!)
Library target will be exported as Audio target in OpenShot:: namespace
(link with OpenShot::Audio to use). find_package(OpenShotAudio) will
load OpenShotAudioConfig.cmake if it can be found on any of CMake's
standard search paths.

VERSION matching is supported (only the X.Y.Z integer components, a
CMake limitation), and dependencies will be looked up by the config.
If an ALSA target needs to be created, it will be automatically.